+++ image-viewer/extractors/imagevenue.js
@@ -4,25 +4,22 @@ import { getUrlFromPage } from './helpers'
 /** @typedef {import('../url-extractor').Extractor} Extractor */
 
 /*
+  link:       https://www.imagevenue.com/ME12EX2A
+  thumbnail:  https://cdn-thumbs.imagevenue.com/16/63/f1/ME12EX2A_t.jpg
+  image:      https://cdn-images.imagevenue.com/26/e4/c7/ME12EX2A_o.jpg
+
+  legacy:
   link:       http://img18127.imagevenue.com/img.php?image=55407_horrorvillian_122_783lo.jpg
   thumbnail:  http://img18127.imagevenue.com/loc783/th_55407_horrorvillian_122_783lo.jpg
-  image:      http://img28127.imagevenue.com/aAfkjfp01fo1i-3407/loc783/55407_horrorvillian_122_783lo.jpg
+  image:      https://cdno-data.imagevenue.com/html.img8127/upload2328/loc783/55407_horrorvillian_122_783lo.jpg
 */
 
 /** @type {Extractor} */
-export const imagevenue = {
+export const imagevenueLegacy = {
   name: 'ImageVenue.com',
-  linkRegEx: new RegExp('imagevenue.com/img.php'),
-  imageUrlRegEx: /id=("|')thepic\1.*src=\1(?<url>[^']*)/i,
+  // Exclude image on https://www.imagevenue.com/
+  linkRegEx: new RegExp('(imagevenue.com/img.php|www.imagevenue.com/\\w+$)'),
+  imageUrlRegEx: /data-toggle="full">\W*<img src="(?<url>[^"]*)/im,
 
-  async getUrl(link, extractor) {
-    const imageUrl = await getUrlFromPage(link, extractor)
-    const pageUrl = link.url
-
-    const url = new URL(pageUrl)
-    url.search = ''
-    url.pathname = imageUrl
-
-    return url.href
-  },
+  getUrl: getUrlFromPage,
 }